backend/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
anthropic
pydantic
motor
httpx
web3
